cmd/groundstation: Use same femto for status submission.
